refactor: structure folders of v0 and v1
