+++ docs/docs-next/pages/index.tsx
@@ -7,6 +7,30 @@ const breakpoints = {
   md: '768px',
   lg: '1080px',
   xl: '1280px',
+} as const
+
+const utilityValues = {
+  value: {
+    base: 'base',
+    xs: 'xs',
+    sm: 'sm',
+    md: 'md',
+    lg: 'lg',
+    xl: 'xl',
+  } as const,
+  fontSize: {
+    base: '12px',
+    md: '24px',
+  } as const,
+  color: {
+    base: 'red',
+    '600px': 'blue',
+    lg: 'green',
+  } as const,
+  isMobileView: {
+    base: true,
+    md: false,
+  } as const,
 }
 
 const useVx = createResponsiveValues({
@@ -14,7 +38,26 @@ const useVx = createResponsiveValues({
   media: 'min',
 })
 
-export default function About() {
-  const isMobileView = useVx({ base: 'mobile', md: 'desktop' })
-  return isMobileView === 'mobile' ? <h3>Mobile View</h3> : <h1>Desktop View</h1>
+export function Component() {
+  const { fontSize, color, value } = useVx(utilityValues)
+  return (
+    <div>
+      <h1 style={{ fontSize, color }}>Hello World: {value}</h1>
+    </div>
+  )
+}
+
+export default function Index() {
+  const { isMobileView } = useVx(utilityValues)
+  return isMobileView ? (
+    <>
+      <Component />
+      <h3>Mobile View</h3>
+    </>
+  ) : (
+    <>
+      <Component />
+      <h1>Desktop View</h1>
+    </>
+  )
 }
