+++ docs/docs-react/src/app/app.tsx
@@ -1,8 +1,7 @@
 import { Route, Routes, Link } from 'react-router-dom'
 import Home from './home'
-import About from './about'
-import TestField from './__test__'
-import V1 from './v1'
+import TestFieldV0 from './__test-v0__'
+import TestFieldV1 from './__test-v1__'
 
 export default function App() {
   return (
@@ -17,29 +16,20 @@ export default function App() {
         }
       />
       <Route
-        path='/about'
+        path='/__test-v0__'
         element={
           <div>
-            <Link to='/'>To Home</Link>
-            <About />
+            <Link to='/'>Home</Link>
+            <TestFieldV0 />
           </div>
         }
       />
       <Route
-        path='/__test__'
+        path='/__test-v1__'
         element={
           <div>
             <Link to='/'>Home</Link>
-            <TestField />
-          </div>
-        }
-      />
-      <Route
-        path='/v1'
-        element={
-          <div>
-            <Link to='/'>Home</Link>
-            <V1 />
+            <TestFieldV1 />
           </div>
         }
       />
